+++ Game/assets/prefabs/CrystalPlatform.prefab.meta
@@ -1,5 +1,5 @@
-asset_hash: 12929283693823254457
+asset_hash: 10176981802147173967
 resources:
   - resource_id: 15301706021286368844
     resource_type: 13
-    resource_hash: 12929283693823254457
+    resource_hash: 10176981802147173967
