+++ Game/assets/scenes/Level.scene.meta
@@ -1,8 +1,8 @@
-asset_hash: 13252677644643378736
+asset_hash: 15809656240647382593
 resources:
   - resource_id: 3947210229329588811
     resource_type: 4
-    resource_hash: 13252677644643378736
+    resource_hash: 15809656240647382593
   - resource_id: 5295014046551349010
     resource_type: 12
     resource_hash: 12219525622439681406
