+++ Gameplay/Gameplay.vcxproj.filters
@@ -517,7 +517,7 @@
     <ClInclude Include="ui\DebugManager.h" />
     <ClInclude Include="misc\AudioManager.h" />
     <ClInclude Include="misc\PillarCheckpoint.h" />
-    <ClInclude Include="CrystalPlatform.h" />
+    <ClInclude Include="misc\CrystalPlatform.h" />
     <ClInclude Include="misc\Spawner.h" />
     <ClInclude Include="misc\DoorController.h" />
     <ClInclude Include="misc\GauntletManager.h" />
@@ -796,7 +796,7 @@
     <ClCompile Include="misc\FancyLights.cpp" />
     <ClCompile Include="misc\AudioManager.cpp" />
     <ClCompile Include="misc\PillarCheckpoint.cpp" />
-    <ClCompile Include="CrystalPlatform.cpp" />
+    <ClCompile Include="misc\CrystalPlatform.cpp" />
     <ClCompile Include="misc\Spawner.cpp" />
     <ClCompile Include="misc\DoorController.cpp" />
     <ClCompile Include="misc\GauntletManager.cpp" />
